test-requirements.txt
daemon
backports-abc
certifi
chardet
croniter
gitdb2
gitpython
gprof2dot
idna
markupsafe
mock
msgpack
ntplib
psutil
pyinotify
pytest
pytest-cov
pytest-html
ansi2html
pytest-pythonpath
python-dateutil
pyyaml
requests
pylint
saltpylint
singledispatch
smmap2
tornado
urllib3
vulners
pyopenssl
pytest-pythonpath
